feat: refactor to use fs promises
